+++ src/FunFair.CodeAnalysis/Helpers/Rules.cs
@@ -23,5 +23,6 @@ namespace FunFair.CodeAnalysis.Helpers
         public const string RuleLoggerParametersShouldBeCalledLogger = @"FFS0019";
         public const string RuleParametersShouldBeInOrder = @"FFS0020";
         public const string RuleDontUseSubstituteReceivedWithZeroNumberOfCalls = @"FFS0021";
+        public const string RuleDontConfigureNullableInCode = @"FFS0022";
     }
 }